# docs/api/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    Multi-tenant CRM for Indian distributors, traders and small manufacturers. Scope stops at the Order: no invoicing, stock or ledger.

    Routes under /api/** require a bearer JWT, including GET /api/v1/auth/me. The exceptions are the four unauthenticated auth calls (signup, login, refresh, logout) and the invitation preview/accept pair; each is marked with an empty security requirement below. GET /public/q/{token} is a separate, unauthenticated route that is not under /api/** at all — the share token itself is the credential. Money is carried as a JSON string, never a number. Errors share one envelope: {"error":{"code","message","fields","fieldCodes"}}; clients translate fieldCodes (docs/api/error-codes.md). The refresh token is never in a body: signup, login, invitation accept and refresh set it as the httpOnly easycrm_rt cookie, and refresh and logout read it.
  title: EasyCRM API
  version: 0.0.1-SNAPSHOT
servers:
- description: Configured deployment origin (easycrm.public-base-url)
  url: http://localhost:8080
security:
- bearer-jwt: []
paths:
...
  /api/v1/auth/invitations/{token}:
    get:
      operationId: preview
      parameters:
      - in: path
        name: token
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationPreviewResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      security: []
      tags:
      - public-invitation-controller
  /api/v1/auth/invitations/{token}/accept:
    post:
      operationId: accept_1
      parameters:
      - in: path
        name: token
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      security: []
      tags:
      - public-invitation-controller
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      security: []
      tags:
      - auth-controller
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses:
        "200":
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      security: []
      tags:
      - auth-controller
  /api/v1/auth/me:
    get:
      operationId: me
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      tags:
      - auth-controller
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      security: []
      tags:
      - auth-controller
  /api/v1/auth/signup:
    post:
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      security: []
      tags:
      - auth-controller
...
  /api/v1/invitations/{id}:
    delete:
      operationId: revoke
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "missing, expired or invalid bearer token"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "authenticated, but the role does not permit this operation"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "the resource does not exist, or belongs to another tenant"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: the request conflicts with existing data
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: semantically invalid request; `fields` names the offending
            inputs
      tags:
      - invitation-controller
...
components:
  schemas:
    AcceptInvitationRequest:
      type: object
      properties:
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        phone:
          type: string
      required:
      - password
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        fieldCodes:
          type: object
          additionalProperties:
            type: string
        fields:
          type: object
          additionalProperties: {}
        message:
          type: string
    ApiErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        email:
          type: string
        role:
          type: string
        tenantId:
          type: string
          format: uuid
        tenantSlug:
          type: string
        userId:
          type: string
          format: uuid
...
    InvitationPreviewResponse:
      type: object
      properties:
        businessName:
          type: string
        email:
          type: string
        role:
          type: string
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
      required:
      - email
      - password
      - slug
...
    MeResponse:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
        tenantId:
          type: string
          format: uuid
        tenantSlug:
          type: string
        userId:
          type: string
          format: uuid
...
    SignupRequest:
      type: object
      properties:
        businessName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        gstin:
          type: string
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
        phone:
          type: string
        slug:
          type: string
          minLength: 1
          pattern: "[a-z0-9-]{3,64}"
        stateCode:
          type: string
          minLength: 1
          pattern: "\\d{2}"
      required:
      - businessName
      - email
      - password
      - slug
      - stateCode
...
  securitySchemes:
    bearer-jwt:
      bearerFormat: JWT
      scheme: bearer
      type: http